优化服务可见性插件
