+++ sermant-agentcore/sermant-agentcore-config/config/config.properties
@@ -12,7 +12,6 @@ agent.config.serviceInjectList=com.huawei.discovery.service.lb.filter.NopInstanc
 agent.service.heartbeat.enable=false
 agent.service.gateway.enable=false
 agent.service.tracing.enable=false
-agent.service.visibility.enable=false
 agent.service.inject.enable=true
 agent.service.dynamic.config.enable=true
