+++ sermant-agentcore/sermant-agentcore-config/config/test/config.properties
@@ -12,7 +12,6 @@ agent.config.serviceInjectList=com.huawei.discovery.service.lb.filter.NopInstanc
 agent.service.heartbeat.enable=true
 agent.service.gateway.enable=true
 agent.service.tracing.enable=true
-agent.service.visibility.enable=true
 agent.service.inject.enable=true
 agent.service.dynamic.config.enable=true
