--- sermant-agentcore/sermant-agentcore-core/src/main/java/com/huaweicloud/sermant/core/service/visibility/api/VisibilityService.java
@@ -1,32 +1,0 @@
-/*
- * Copyright (C) 2022-2022 Huawei Technologies Co., Ltd. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.huaweicloud.sermant.core.service.visibility.api;
-
-import com.huaweicloud.sermant.core.service.BaseService;
-
-/**
- * 服务下线接口处理
- *
- * @author zhp
- * @since 2022-12-10
- */
-public interface VisibilityService extends BaseService {
-    /**
-     * 服务重连处理
-     */
-    void reconnectHandler();
-}
