+++ sermant-plugins/sermant-service-visibility/service-visibility-plugin/src/main/java/com/huaweicloud/visibility/interceptor/SpringCloudMappingRegistryInterceptor.java
@@ -18,12 +18,12 @@
 package com.huaweicloud.visibility.interceptor;
 
 import com.huaweicloud.sermant.core.plugin.agent.entity.ExecuteContext;
-import com.huaweicloud.sermant.core.service.visibility.common.CollectorCache;
-import com.huaweicloud.sermant.core.service.visibility.common.OperateType;
-import com.huaweicloud.sermant.core.service.visibility.common.ServiceType;
-import com.huaweicloud.sermant.core.service.visibility.entity.Contract;
-import com.huaweicloud.sermant.core.service.visibility.entity.MethodInfo;
-import com.huaweicloud.sermant.core.service.visibility.entity.ServerInfo;
+import com.huaweicloud.visibility.common.CollectorCache;
+import com.huaweicloud.visibility.common.OperateType;
+import com.huaweicloud.visibility.common.ServiceType;
+import com.huaweicloud.visibility.entity.Contract;
+import com.huaweicloud.visibility.entity.MethodInfo;
+import com.huaweicloud.visibility.entity.ServerInfo;
 
 import org.springframework.web.servlet.mvc.method.RequestMappingInfo;
