+++ sermant-plugins/sermant-service-visibility/service-visibility-plugin/src/main/java/com/huaweicloud/visibility/interceptor/SpringCloudReactiveDiscoveryClientInterceptor.java
@@ -18,10 +18,10 @@
 package com.huaweicloud.visibility.interceptor;
 
 import com.huaweicloud.sermant.core.plugin.agent.entity.ExecuteContext;
-import com.huaweicloud.sermant.core.service.visibility.common.CollectorCache;
-import com.huaweicloud.sermant.core.service.visibility.common.OperateType;
-import com.huaweicloud.sermant.core.service.visibility.entity.Consanguinity;
-import com.huaweicloud.sermant.core.service.visibility.entity.ServerInfo;
+import com.huaweicloud.visibility.common.CollectorCache;
+import com.huaweicloud.visibility.common.OperateType;
+import com.huaweicloud.visibility.entity.Consanguinity;
+import com.huaweicloud.visibility.entity.ServerInfo;
 
 import reactor.core.publisher.Flux;
