+++ sermant-plugins/sermant-service-visibility/service-visibility-plugin/src/test/java/com/huaweicloud/visibility/common/CollectorCacheTest.java
@@ -1,26 +1,8 @@
-/*
- * Copyright (C) 2022-2022 Huawei Technologies Co., Ltd. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- *
- */
+package com.huaweicloud.visibility.common;
 
-package com.huaweicloud.sermant.core.service.visibility.common;
-
-import com.huaweicloud.sermant.core.service.visibility.entity.Consanguinity;
-import com.huaweicloud.sermant.core.service.visibility.entity.Contract;
-import com.huaweicloud.sermant.core.service.visibility.entity.MethodInfo;
-
+import com.huaweicloud.visibility.entity.Consanguinity;
+import com.huaweicloud.visibility.entity.Contract;
+import com.huaweicloud.visibility.entity.MethodInfo;
 import org.junit.Assert;
 import org.junit.Test;
 
@@ -28,12 +10,6 @@ import java.lang.reflect.Method;
 import java.util.ArrayList;
 import java.util.List;
 
-/**
- * 服务可见性缓存测试类
- *
- * @author zhp
- * @since 2022-12-09
- */
 public class CollectorCacheTest {
     private static final String DEFAULT_IP = "127.0.0.1";
